Removed `exec` calls
I originally wanted to have all the "paramX" and "stringX" fields as seperate attributes, but this is not only ugly but also unnecessary.

I merged them into a single list residing in the "param" or "string" attribute respectively.

I also included a bit of typing here, since that will be a topic of future commits regardless.