Second housekeeping pass
Did a slight cosmetics and documentation pass on the object model code, similar to the one done on the core.

It was not nearly as thorough as I still plan on rewriting this code to use less classes and fewer attributes, but for the time being, this will have to do.