Update query command syntax
Altered the query command syntax to be more user-friendly. This now uses an internal wrapper to avoid having to have identical attributes and setter methods for each command.

The actual value is available via the CallableAttribute.value attribute.

CallableAttribute.__call__ currently requires that CallableAttribute.__get__ has been called just before, which it uses to store the relation to its containing object. This is fragile and prone to race conditions, this should be replaced with a better system.