Move expo_scaled factory to utilities
This also made expo_scaled a configurable factory, rather than just using a hard-coded scaling factor.

This commit will change backup times as the previous implementation was scaling the generated value twice. This might need more tweaking if retry times are too great, I was not able to test this due to the API being rock solid this afternoon.