New docstring syntax
This syntax is a sphinx-heavy style that excludes type annotations. This hybrid is currently the only way to get reasonably elegant HTML docs without losing access to autodoc.

Napoleon and other type hint introspection utilities lack the flexibility to use canonical types in the docs, which makes them difficult to read and work with.